my_mcp_server/pyproject.toml
[project]
name = "my-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.12.4",
    "pillow>=11.3.0",
]
